Alert implemented, cupon fix, new styles

+++ src/components/ListOwnOffers/ListOwnOffers.js
@@ -5,8 +5,8 @@ import { FaTrashAlt } from "react-icons/fa";
 import { v4 as uuidv4 } from 'uuid';
 import { ApiContext } from '../../helper/context/ApiContext';
 import { ItemsContext } from '../../helper/context/ItemsContext';
-import toast from "react-hot-toast";
 import { ModalEditOffer } from '../ModalEditOffer/ModalEditOffer';
+import Swal from 'sweetalert2'
 
 export function ListOwnOffers() {
     const { fetchDataOffers } = useContext(ApiContext);
@@ -14,22 +14,39 @@ export function ListOwnOffers() {
     const [myOffers, setMyOffers] = useState(offers.filter(e => e.idUsers === user.id));
     const [interim, setInterim] = useState();
 
-    
-    useEffect(()=> {
+
+    useEffect(() => {
         setMyOffers(offers.filter(e => e.idUsers === user.id))
     }, [interim])
 
 
     const removeOffer = (product, id) => {
-        let interim = myOffers.find((offer, indice) => indice === id);
-        let offers_prov = myOffers.filter((offer, indice) => indice !== id);
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                let interim = myOffers.find((offer, indice) => indice === id);
+                let offers_prov = myOffers.filter((offer, indice) => indice !== id);
 
-        fetch(`http://localhost:4000/offers/${interim.id}`, {
-            method: 'DELETE'
-        }).then(() => setMyOffers(offers_prov))
-            .then(() => fetchDataOffers())
-            .then(() => toast('Deleted!', { icon: '🗑️' }))
-            .catch(error => console.log(error));
+                fetch(`http://localhost:4000/offers/${interim.id}`, {
+                    method: 'DELETE'
+                }).then(() => setMyOffers(offers_prov))
+                    .then(() => fetchDataOffers())
+                    .then(() => Swal.fire(
+                        'Deleted!',
+                        'Your file has been deleted.',
+                        'success'
+                    ))
+                        .catch(error => console.log(error));
+            }
+        })
+
     }
 
     return (
@@ -49,7 +66,7 @@ export function ListOwnOffers() {
 
                             {/* //TODO CAMBIAR BOTON PARA QUE HAGA UN METHOD DELETE */}
                             <div className="pointer">
-                                <ModalEditOffer item={item} index={index} setInterim={setInterim}/>
+                                <ModalEditOffer item={item} index={index} setInterim={setInterim} />
                                 <Button onClick={e => { removeOffer(item, index) }} variant="outline-danger" ><FaTrashAlt /></Button>
                             </div>
